Updated conform reviews
